fixed unittests

+++ package/tests/test_downloader.py
@@ -31,8 +31,10 @@ class TestDownloader(unittest.TestCase):
         self.dest_path = os.path.join(self.data_dir, 'test_download_dest')
 
     def tearDown(self):
-        shutil.rmtree(self.tmp_path)
-        shutil.rmtree(self.dest_path)
+        if os.path.isdir(self.tmp_path):
+            shutil.rmtree(self.tmp_path)
+        if os.path.isdir(self.dest_path):
+            shutil.rmtree(self.dest_path)
 
     @skip_if_import_exception
     def test_download(self):
